.github/bundle-application/requirements.txt
awsebcli==3.18.2
pytest==6.0.1
boto3==1.14.37
